kakaochannelSplit
