+++ build.gradle
@@ -51,3 +51,36 @@ subprojects {
 		useJUnitPlatform()
 	}
 }
+//
+project(":trackingpost-core") {
+	// 공통 코드 모듈
+	bootJar { enabled = false } // core 은 bootJar 로 패키징 할 필요 없음
+	jar { enabled = true }
+
+	dependencies {
+	}
+}
+project(":servicehome") {
+	bootJar { enabled = true }
+	jar { enabled = false }
+
+	dependencies {
+		implementation project(':trackingpost-core') // 컴파일 시 trackingpost-core project 로딩
+	}
+}
+project(":externalpost") {
+	bootJar { enabled = true }
+	jar { enabled = false }
+
+	dependencies {
+		implementation project(':trackingpost-core') // 컴파일 시 trackingpost-core project 로딩
+	}
+}
+project(":kakaochannel") {
+	bootJar { enabled = true }
+	jar { enabled = false }
+
+	dependencies {
+		implementation project(':trackingpost-core') // 컴파일 시 trackingpost-core project 로딩
+	}
+}
