+++ kakaochannel/src/main/resources/application.properties
@@ -1,1 +1,2 @@
-
+#server port
+server.port = 8082
