--- servicehome/src/main/java/com/lostcatbox/trackingpost/service/KindRequest.java
@@ -1,5 +1,0 @@
-package com.lostcatbox.trackingpost.service;
-
-public enum KindRequest {
-    Kakao,Web
-}
