+++ servicehome/src/main/java/com/lostcatbox/trackingpost/service/provider/PostProvider.java
@@ -1,7 +1,7 @@
 package com.lostcatbox.trackingpost.service.provider;
 
-import com.lostcatbox.trackingpost.domain.PostDto;
-import com.lostcatbox.trackingpost.service.PostCompanyEnum;
+import com.example.trackingpostcore.domain.PostDto;
+import com.example.trackingpostcore.domain.PostCompanyEnum;
 
 public interface PostProvider {
