+++ settings.gradle
@@ -2,5 +2,6 @@ rootProject.name = 'trackingpost'
 include(
         "servicehome",
         "externalpost",
-        "kakaochannel"
+        "kakaochannel",
+        "trackingpost-core"
 )
