+++ trackingpost-core/src/main/java/com/example/trackingpostcore/domain/Post.java
@@ -1,4 +1,4 @@
-package com.lostcatbox.trackingpost.domain;
+package com.example.trackingpostcore.domain;
 
 import lombok.AccessLevel;
 import lombok.Builder;
