+++ trackingpost-core/src/main/java/com/example/trackingpostcore/domain/PostCompanyEnum.java
@@ -1,0 +1,5 @@
+package com.example.trackingpostcore.domain;
+
+public enum PostCompanyEnum {
+    CJ,CU
+}
